+++ sql/Data.sql
@@ -231,7 +231,7 @@ DROP TABLE IF EXISTS `formation`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `formation` (
-  `id` mediumint(3) unsigned NOT NULL,
+  `id` tinyint(3) unsigned NOT NULL,
   `name` varchar(255) NOT NULL,
   `grid1` tinyint(3) NOT NULL,
   `prop1` int(10) NOT NULL,
